pyproject.toml
[project]
name = "moa-debate"
version = "0.1.0"
description = "Multi-model AI debate system using Mixture-of-Agents and Expert Panel patterns"
requires-python = ">=3.9"
dependencies = [
    "litellm>=1.40.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[project.scripts]
moa = "moa.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
